requirements.txt
streamlit
langchain
langchain-core
langchain-openai
langchain-qdrant
qdrant-client
pypdf
python-dotenv